+++ saleor/graphql/product/resolvers.py
@@ -17,24 +17,38 @@ COLLECTION_SEARCH_FIELDS = ('name', 'slug')
 ATTRIBUTES_SEARCH_FIELDS = ('name', 'slug')
 
 
-def resolve_attributes(info, category_id, query):
+def _filter_attributes_by_product_types(attribute_qs, product_qs):
+    product_types = set(product_qs.values_list('product_type_id', flat=True))
+    return attribute_qs.filter(
+        Q(product_type__in=product_types)
+        | Q(product_variant_type__in=product_types))
+
+
+def resolve_attributes(info, category_id=None, collection_id=None, query=None):
     qs = models.Attribute.objects.all()
     qs = filter_by_query_param(qs, query, ATTRIBUTES_SEARCH_FIELDS)
+
     if category_id:
-        # Get attributes that are used with product types
-        # within the given category.
+        # Filter attributes by product types belonging to the given category.
         category = graphene.Node.get_node_from_global_id(
             info, category_id, Category)
-        if category is None:
-            return qs.none()
-        tree = category.get_descendants(include_self=True)
-        product_types = {
-            obj[0]
-            for obj in models.Product.objects.filter(
-                category__in=tree).values_list('product_type_id')}
-        qs = qs.filter(
-            Q(product_type__in=product_types)
-            | Q(product_variant_type__in=product_types))
+        if category:
+            tree = category.get_descendants(include_self=True)
+            product_qs = models.Product.objects.filter(category__in=tree)
+            qs = _filter_attributes_by_product_types(qs, product_qs)
+        else:
+            qs = qs.none()
+
+    if collection_id:
+        # Filter attributes by product types belonging to the given collection.
+        collection = graphene.Node.get_node_from_global_id(
+            info, collection_id, Collection)
+        if collection:
+            product_qs = collection.products.all()
+            qs = _filter_attributes_by_product_types(qs, product_qs)
+        else:
+            qs = qs.none()
+
     qs = qs.order_by('name')
     qs = qs.distinct()
     return gql_optimizer.query(qs, info)